added: cart/kept handling

+++ src/main/java/com/example/MyBookShopApp/controller/BookShopPostponed.java
@@ -29,7 +29,7 @@ public class BookShopPostponed {
         Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         try {
             if (((UserEntityDetails) principal).getUsername() != null && !((UserEntityDetails) principal).getUsername().equals("")) {
-                return bookService.getPostponedCount();
+                return bookService.getPostponedCount(principal);
             }
         } catch (Exception e) {
             return bookService.getPostponedCountTempUser(postponedContents);
@@ -42,7 +42,7 @@ public class BookShopPostponed {
         Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         try {
             if (((UserEntityDetails) principal).getUsername() != null && !((UserEntityDetails) principal).getUsername().equals("")) {
-                return bookService.getCartCount();
+                return bookService.getCartCount(principal);
             }
         } catch (Exception e) {
             return bookService.getCartCountTempUser(cartContents);
@@ -59,31 +59,81 @@ public class BookShopPostponed {
     @GetMapping("/postponed")
     public String handlePostponedRequest(@CookieValue(name = "postponedContents", required = false) String postponedContents,
                                          Model model) {
-        if (bookService.getPostponedCount() == 0) {
-            model.addAttribute("isPostponedEmpty", true);
-        } else {
-            model.addAttribute("isPostponedEmpty", false);
-            model.addAttribute("bookPostponed", bookService.getBookListInPostponed());
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        try {
+            if (((UserEntityDetails) principal).getUsername() != null && !((UserEntityDetails) principal).getUsername().equals("")) {
+                if (bookService.getPostponedCount(principal) == 0) {
+                    model.addAttribute("isPostponedEmpty", true);
+                } else {
+                    model.addAttribute("isPostponedEmpty", false);
+                    model.addAttribute("bookPostponed", bookService.getBookListInPostponed(principal));
+                    model.addAttribute("postponedSize", bookService.getPostponedCount(principal));
+                    model.addAttribute("cartContentsSize", bookService.getCartCount(principal));
+                }
+            }
+        } catch (Exception e) {
+            if (bookService.getPostponedCountTempUser(postponedContents) == 0) {
+                model.addAttribute("isPostponedEmpty", true);
+            } else {
+                model.addAttribute("isPostponedEmpty", false);
+                model.addAttribute("bookPostponed", bookService.getBookListInPostponedUserTemp(postponedContents));
+            }
         }
         return "postponed";
     }
 
     @PostMapping("/changeBookStatus/postponed/remove/{slug}")
     public String handleRemoveBookFromPostponedRequest(@PathVariable("slug") String slug,
+                                                       @CookieValue(name = "cartContents", required = false) String cartContents,
+                                                       @CookieValue(name = "postponedContents", required = false) String postponedContents,
+                                                       HttpServletResponse response,
                                                        Model model) throws BookstoreAPiWrongParameterException {
         Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        bookService.removePostponedItem(slug, principal);
-        model.addAttribute("postponedSize", bookService.getPostponedCount());
-        model.addAttribute("cartContentsSize", bookService.getCartCount());
+        try {
+            if (((UserEntityDetails) principal).getUsername() != null && !((UserEntityDetails) principal).getUsername().equals("")) {
+                bookService.removePostponedItem(slug, principal);
+                model.addAttribute("postponedSize", bookService.getPostponedCount(principal));
+                model.addAttribute("cartContentsSize", bookService.getCartCount(principal));
+                model.addAttribute("bookPostponed", bookService.getBookListInPostponed(principal));
+            }
+        } catch (Exception e) {
+            Cookie cookie = bookService.removePostponedItemTempUser(slug, postponedContents);
+            response.addCookie(cookie);
+            model.addAttribute("postponedSize", bookService.getPostponedCountTempUser(postponedContents));
+            model.addAttribute("isPostponedEmpty", bookService.getPostponedCountTempUser(postponedContents) == 0);
+            model.addAttribute("bookPostponed", bookService.getBookListInPostponedUserTemp(postponedContents));
+        }
         return "redirect:/books/postponed";
     }
 
     @PostMapping("/changeBookStatus/postponed/buy/{slug}")
     public String handleBuyBookFromPostponedRequest(@PathVariable("slug") String slug,
+                                                    @CookieValue(name = "cartContents", required = false) String cartContents,
+                                                    @CookieValue(name = "postponedContents", required = false) String postponedContents,
+                                                    HttpServletResponse response,
                                                     Model model) {
-        bookService.addCartItem(slug);
-        model.addAttribute("postponedSize", bookService.getPostponedCount());
-        model.addAttribute("cartContentsSize", bookService.getCartCount());
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        try {
+            if (((UserEntityDetails) principal).getUsername() != null && !((UserEntityDetails) principal).getUsername().equals("")) {
+                bookService.addCartItem(slug, principal);
+                model.addAttribute("postponedSize", bookService.getPostponedCount(principal));
+                model.addAttribute("cartContentsSize", bookService.getCartCount(principal));
+                model.addAttribute("bookPostponed", bookService.getBookListInPostponed(principal));
+            }
+        } catch (Exception e) {
+            Cookie cookie = bookService.addCartItemTempUser(slug, cartContents);
+            response.addCookie(cookie);
+            model.addAttribute("isCartEmpty", false);
+
+            Cookie cookiePostponed = bookService.removePostponedItemTempUser(slug, postponedContents);
+            response.addCookie(cookiePostponed);
+
+            model.addAttribute("postponedSize", bookService.getPostponedCountTempUser(postponedContents));
+            model.addAttribute("cartContentsSize", bookService.getCartCountTempUser(cartContents));
+            model.addAttribute("bookPostponed", bookService.getBookListInPostponedUserTemp(postponedContents));
+        }
         return "redirect:/books/postponed";
     }
 
@@ -98,8 +148,8 @@ public class BookShopPostponed {
         try {
             if (((UserEntityDetails) principal).getUsername() != null && !((UserEntityDetails) principal).getUsername().equals("")) {
                 bookService.addPostponedItem(slug, principal);
-                model.addAttribute("postponedSize", bookService.getPostponedCount());
-                model.addAttribute("cartContentsSize", bookService.getCartCount());
+                model.addAttribute("postponedSize", bookService.getPostponedCount(principal));
+                model.addAttribute("cartContentsSize", bookService.getCartCount(principal));
             }
         } catch (Exception e) {
             Cookie cookie = bookService.addPostponedItemTempUser(slug, postponedContents);
